rename thd_sm100.py to thd_helpers.py
