+++ python/cudnn/sdpa/fwd/kernels/prefill_d128_f16_sm100.py
@@ -37,10 +37,10 @@ torch fp32 reference (atol 0.025) across single-tile, multi-tile GQA, and the
 
 THD / varlen (``CFG.THD_VARLEN=1``): packed ``[1,T,H,D]`` Q/K/V + ``cu_seqlens``
 coord offset (applied to BOTH Q slabs under TILES_Q=2), per-batch O
-TMA-descriptor array (shared ``thd_sm100.py``), packed ragged-Stats LSE
+TMA-descriptor array (shared ``thd_helpers.py``), packed ragged-Stats LSE
 (head-major ``[1,QH,head_stride]`` or token-major ``[T,QH]`` — the epilogue
 branches on the LSE tensor's static rank) — via
-the shared ``_common_sm100`` / ``thd_sm100`` mechanism (same as the SM100 qwen
+the shared ``_common_sm100`` / ``thd_helpers`` mechanism (same as the SM100 qwen
 / dsv4 kernels).  The dense ``[B,S,H,D]`` path is byte-identical (folds out at
 ``THD_VARLEN=0``).
 
@@ -224,7 +224,7 @@ _resolve_seqlen_q = _sdpa_h.resolve_seqlen_q
 # seq_kv_lens overloaded as the THD metadata buffer (int32 len 4B+4):
 #   [0..B-1]=seq_kv_lens  [B..2B]=cu_q(B+1)  [2B+1..3B+1]=cu_k(B+1)
 #   [3B+2..4B+1]=batch_remap(B)  [4B+2]=live units  [4B+3]=claim counter
-from cudnn.sdpa.fwd.kernels.thd_sm100 import build_thd_meta_o_descs_kernel as _build_thd_meta_o_descs_kernel, TENSOR_MAP_QWORDS, THD_SETUP_THREADS
+from cudnn.sdpa.fwd.kernels.thd_helpers import build_thd_meta_o_descs_kernel as _build_thd_meta_o_descs_kernel, TENSOR_MAP_QWORDS, THD_SETUP_THREADS
 
 _TENSOR_MAP_QWORDS = TENSOR_MAP_QWORDS
 # The setup kernel builds the THD metadata buffer DEVICE-side from the
